.
Given an array of integers, find all pairs of elements whose sum is equal to a given target value.

Input:

An array of integers: arr

Target value: target

Output:

A list of pairs, where each pair consists of two integers from the array whose sum is equal to the target value.

Input:

arr = [2, 4, 6, 3, 9, 1]

target = 7



Output:

[(2, 5), (4, 3)]



Explanation:

In the given example, we need to find pairs whose sum is equal to 7. The pairs that satisfy this condition are (2, 5) and (4, 3). The pair (2, 5) is formed by adding the first element (2) with the fourth element (5) in the array. Similarly, the pair (4, 3) is formed by adding the second element (4) with the third element (3) in the array.